lesson-04-memory/pyproject.toml
[tool.poetry]
name = "lesson-04-memory"
version = "0.1.0"
description = "LangChain Lesson 4: Memory & Conversation Management"
authors = ["LangChain Course"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.3.0"
langchain-openai = "^0.3.3"
langchain-anthropic = "^0.3.0"
langchain-ollama = "^0.3.0"
langchain-google-genai = "^2.1.0"
langchain-deepseek = "^0.1.0"
langchain-community = "^0.3.0"
langchain-core = "^0.3.33"
python-dotenv = "^1.0.0"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"

# Memory & Vector Store dependencies
chromadb = "^0.4.15"
faiss-cpu = "^1.7.4"
sentence-transformers = "^2.2.2"
tiktoken = "^0.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 